Figure_To_Pdf: use max plane size to select the appropriate zoom level
The logic selecting the best resolution for the export can select a
plane as large as omero.client.download_as.max_size. With a default value
of  144000000, this means regions as wide as 12k x 12k can be requested
to the server.

This commit proposes to use the max plane width and height instead to
select the appropriate resolution level. Both configurations default to
3k and are already used for identifying an image as "big" and navigating
through the resolution levels.